\name{aamatch-package}
\alias{aamatch-package}
\alias{aamatch}
\docType{package}
\title{
\packageTitle{aamatch}
}
\description{
\packageDescription{aamatch}
}
\details{
Package aamatch implements a simple version of multivariate matching in observational studies, using propensity scores, minimum distance matching, near-exact matching and fine balance.  The main function in the package is
artlessV2().  artlessV2() calls the function alittleArt(), and the latter gives the user finer control over the match.
}
\author{
\packageAuthor{aamatch}

Maintainer: \packageMaintainer{aamatch}
}
\references{
Rosenbaum, P. R. (2020a) <doi:10.1007/978-3-030-46405-9> Design of Observational Studies (2nd Edition). New York: Springer.

Rosenbaum, P. R. (2020b). <doi:10.1146/annurev-statistics-031219-041058> Modern algorithms for matching in observational studies. Annual Review of Statistics and Its Application, 7(1), 143-176.

Rosenbaum, P. R. (2025) <doi:10.1007/978-3-031-90494-3> Introduction to the Theory of Observational Studies.  New York: Springer.

Zhang, B., D. S. Small, K. B. Lasater, M. McHugh, J. H. Silber, and P. R. Rosenbaum (2023) <doi:10.1080/01621459.2021.1981337> Matching one sample according to two criteria in observational studies. Journal of the American Statistical Association, 118, 1140-1151.
}
\keyword{package}




